requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-cloud-speech
moviepy
pyaudio
